# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "SystemStatusResponse": {
        "description": "System status response model.",
        "properties": {
          "deployment_info": {
            "additionalProperties": true,
            "default": {},
            "title": "Deployment Info",
            "type": "object"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms"
          },
          "services": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Services",
            "type": "object"
          },
          "status": {
            "description": "healthy, degraded, critical",
            "title": "Status",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "uptime_seconds": {
            "title": "Uptime Seconds",
            "type": "number"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "uptime_seconds",
          "services",
          "version"
        ],
        "title": "SystemStatusResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Advanced Near Earth Object detection System - RESTful API for artificial NEO detection",
    "title": "aNEOS API",
    "version": "2.0.0",
    "x-logo": {
      "url": "https://example.com/aneos-logo.png"
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/status": {
      "get": {
        "description": "Get comprehensive system status for administrators.",
        "operationId": "get_system_status_api_v1_admin_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get System Status",
        "tags": [
          "Administration"
        ]
      }
    },
...
    "/api/v1/monitoring/dashboard": {
      "get": {
        "description": "Get comprehensive dashboard data for monitoring interface.",
        "operationId": "get_dashboard_data_api_v1_monitoring_dashboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Dashboard Data Api V1 Monitoring Dashboard Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dashboard Data",
        "tags": [
          "Monitoring"
        ]
      }
    },
...
    "/dashboard/": {
      "get": {
        "description": "Main dashboard page.",
        "operationId": "dashboard_home_dashboard__get",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dashboard Home",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/admin": {
      "get": {
        "description": "Administration dashboard.",
        "operationId": "admin_dashboard_dashboard_admin_get",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Dashboard",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/analysis": {
      "get": {
        "description": "NEO analysis dashboard.",
        "operationId": "analysis_dashboard_dashboard_analysis_get",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analysis Dashboard",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/dashboard/monitoring": {
      "get": {
        "description": "System monitoring dashboard.",
        "operationId": "monitoring_dashboard_dashboard_monitoring_get",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitoring Dashboard",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
  }
}